requirements.txt
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-core==2.3.2
google-cloud-dialogflow==2.16.1
google-cloud-storage==2.5.0
python-dotenv==0.21.0
python-telegram-bot==13.14
vk-api==11.9.9
